# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /books/search:
    get:
      operationId: searchBooks
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookSearchResult"
                required: [results]
  /trending/reviews:
    get:
      operationId: getTrendingReviews
      responses:
        "200":
          description: Trending reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrendingReview"
                required: [reviews]
  /me/books/import-file:
    post:
      operationId: importBookFile
      responses:
        "201":
          description: Imported file-backed book
          content:
            application/json:
              schema:
                type: object
                properties:
                  book:
                    $ref: "#/components/schemas/Book"
                required: [book]
  /me/books/{id}/epub-url:
    get:
      operationId: getBookEpubUrl
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signed EPUB URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required: [url]
  /me/books/{id}/review:
    get:
      operationId: getMyBookReview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current user's review
          content:
            application/json:
              schema:
                type: object
                properties:
                  review:
                    $ref: "#/components/schemas/Review"
                required: [review]
    put:
      operationId: upsertMyBookReview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved review
          content:
            application/json:
              schema:
                type: object
                properties:
                  review:
                    $ref: "#/components/schemas/Review"
                required: [review]
    delete:
      operationId: deleteMyBookReview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Review deleted
  /me/integrations/hardcover:
    get:
      operationId: getHardcoverIntegration
      responses:
        "200":
          description: Hardcover connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                required: [connected]
    delete:
      operationId: disconnectHardcover
      responses:
        "204":
          description: Hardcover disconnected
  /me/integrations/hardcover/connect:
    post:
      operationId: connectHardcover
      responses:
        "200":
          description: Hardcover connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  previewCount:
                    type: integer
                required: [connected, previewCount]
  /me/integrations/hardcover/import:
    post:
      operationId: importHardcoverLibrary
      responses:
        "200":
          description: Hardcover import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  updated:
                    type: integer
                  total:
                    type: integer
                required: [imported, updated, total]
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BookFormat:
      type: string
      enum:
        - Paperback
        - Ebook
        - Audiobook
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
        format:
          $ref: "#/components/schemas/BookFormat"
        source:
          type: string
        demoBookId:
          type: string
          nullable: true
        coverGradient:
          type: array
          items:
            type: string
        visualStyle:
          type: string
        spoilerMode:
          type: string
        currentChapter:
          type: integer
        currentPage:
          type: integer
        currentAudioTimestamp:
          type: string
        progress:
          type: integer
        userNote:
          type: string
        tagline:
          type: string
          nullable: true
        heroImage:
          type: string
          nullable: true
        epubObjectKey:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        lastReadCfi:
          type: string
          nullable: true
        totalChapters:
          type: integer
          nullable: true
        readingStatus:
          type: string
        seriesName:
          type: string
          nullable: true
        seriesOrder:
          type: integer
          nullable: true
      required:
        - id
        - title
        - author
        - format
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        userBookId:
          type: string
          format: uuid
        rating:
          type: integer
        body:
          type: string
          nullable: true
        containsSpoilers:
          type: boolean
        shareToTrending:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userBookId
        - rating
        - containsSpoilers
        - shareToTrending
        - createdAt
        - updatedAt
    TrendingReview:
      allOf:
        - $ref: "#/components/schemas/Review"
        - type: object
          properties:
            bookTitle:
              type: string
            author:
              type: string
            authorHandle:
              type: string
          required:
            - bookTitle
            - author
            - authorHandle
    BookSearchResult:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        author:
          type: string
        firstPublishYear:
          type: integer
          nullable: true
        pageCount:
          type: integer
          nullable: true
        coverUrl:
          type: string
          nullable: true
        coverUrlLarge:
          type: string
          nullable: true
        workKey:
          type: string
        source:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
      required:
        - key
        - title
        - author
        - workKey