# lib/api-zod/src/generated/types/trendingReview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from "./review";

export type TrendingReview = Review & {
  bookTitle: string;
  author: string;
  authorHandle: string;
};